pyproject.toml
[project]
name = "rag-agent-local"
version = "0.1.0"
description = "Local-first RAG + Agent project for interview-ready demonstrations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "python-multipart>=0.0.20",
  "httpx>=0.28.0",
  "chromadb>=1.0.0",
  "numpy>=2.2.0",
  "rank-bm25>=0.2.2",
  "pypdf>=5.0.0",
  "python-docx>=1.1.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "ruff>=0.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
